Renamed ema_control package to emacontrol
